+++ spreadsheet-xmlss/src/main/java/ec/util/spreadsheet/xmlss/XmlssBookWriter.java
@@ -1,35 +1,35 @@
 /*
  * Copyright 2013 National Bank of Belgium
  *
- * Licensed under the EUPL, Version 1.1 or – as soon they will be approved 
+ * Licensed under the EUPL, Version 1.1 or – as soon they will be approved
  * by the European Commission - subsequent versions of the EUPL (the "Licence");
  * You may not use this work except in compliance with the Licence.
  * You may obtain a copy of the Licence at:
  *
  * http://ec.europa.eu/idabc/eupl
  *
- * Unless required by applicable law or agreed to in writing, software 
+ * Unless required by applicable law or agreed to in writing, software
  * distributed under the Licence is distributed on an "AS IS" basis,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Licence for the specific language governing permissions and 
+ * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
 package ec.util.spreadsheet.xmlss;
 
 import ec.util.spreadsheet.Book;
-import ec.util.spreadsheet.Cell;
 import ec.util.spreadsheet.Sheet;
+import org.checkerframework.checker.nullness.qual.NonNull;
 import shaded.spreadsheet.nbbrd.io.xml.Stax;
-import java.io.IOException;
-import java.io.OutputStream;
-import java.nio.charset.Charset;
+
 import javax.xml.stream.XMLOutputFactory;
 import javax.xml.stream.XMLStreamException;
 import javax.xml.stream.XMLStreamWriter;
-import org.checkerframework.checker.nullness.qual.NonNull;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.nio.charset.Charset;
+import java.util.Date;
 
 /**
- *
  * @author Philippe Charles
  */
 final class XmlssBookWriter {
@@ -59,7 +59,7 @@ final class XmlssBookWriter {
 
     private static void write(BasicXmlssWriter f, Book book) throws IOException, XMLStreamException {
         f.beginWorkbook();
-        int sheetCount = book.getSheetCount();
+        int sheetCount = book.getSheetCount2();
         for (int s = 0; s < sheetCount; s++) {
             write(f, book.getSheet(s));
         }
@@ -74,7 +74,7 @@ final class XmlssBookWriter {
         for (int i = 0; i < rowCount; i++) {
             f.beginRow();
             for (int j = 0; j < columnCount; j++) {
-                write(f, sheet.getCell(i, j));
+                writeCellValue(f, sheet.getCellValue(i, j));
             }
             f.endRow();
         }
@@ -82,14 +82,14 @@ final class XmlssBookWriter {
         f.endWorksheet();
     }
 
-    private static void write(BasicXmlssWriter f, Cell cell) throws XMLStreamException {
-        if (cell != null) {
-            if (cell.isDate()) {
-                f.writeCell(cell.getDate());
-            } else if (cell.isNumber()) {
-                f.writeCell(cell.getDouble());
-            } else if (cell.isString()) {
-                f.writeCell(cell.getString());
+    private static void writeCellValue(BasicXmlssWriter f, Object cellValue) throws XMLStreamException {
+        if (cellValue != null) {
+            if (cellValue instanceof Date) {
+                f.writeCell((Date) cellValue);
+            } else if (cellValue instanceof Number) {
+                f.writeCell(((Number) cellValue).doubleValue());
+            } else if (cellValue instanceof String) {
+                f.writeCell((String) cellValue);
             }
         } else {
             f.writeCell();